role permission udpate

+++ app/Modules/User/Controllers/RoleController.php
@@ -45,6 +45,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(!auth()->user()->hasPermission('roles.view'), 403, 'You do not have permission to view roles.');
+        
         $roles = $this->roleService->getAllRoles();
         return view('admin.roles.index', compact('roles'));
     }
@@ -55,6 +57,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        abort_if(!auth()->user()->hasPermission('roles.create'), 403, 'You do not have permission to create roles.');
+        
         $permissions = $this->permissionRepository->getActiveByEnabledModules();
         return view('admin.roles.create', compact('permissions'));
     }
@@ -64,6 +68,8 @@ class RoleController extends Controller
      */
     public function store(StoreRoleRequest $request)
     {
+        abort_if(!auth()->user()->hasPermission('roles.create'), 403, 'You do not have permission to create roles.');
+        
         $result = $this->roleService->createRole($request->validated());
 
         if ($result['success']) {
@@ -83,6 +89,8 @@ class RoleController extends Controller
      */
     public function show(int $id)
     {
+        abort_if(!auth()->user()->hasPermission('roles.view'), 403, 'You do not have permission to view roles.');
+        
         $role = $this->roleService->getRoleById($id);
         
         if (!$role) {
@@ -100,6 +108,8 @@ class RoleController extends Controller
      */
     public function edit(int $id)
     {
+        abort_if(!auth()->user()->hasPermission('roles.edit'), 403, 'You do not have permission to edit roles.');
+        
         $role = $this->roleService->getRoleById($id);
         
         if (!$role) {
@@ -117,6 +127,8 @@ class RoleController extends Controller
      */
     public function update(UpdateRoleRequest $request, int $id)
     {
+        abort_if(!auth()->user()->hasPermission('roles.edit'), 403, 'You do not have permission to edit roles.');
+        
         $result = $this->roleService->updateRole($id, $request->validated());
 
         if ($result['success']) {
@@ -136,6 +148,8 @@ class RoleController extends Controller
      */
     public function destroy(int $id)
     {
+        abort_if(!auth()->user()->hasPermission('roles.delete'), 403, 'You do not have permission to delete roles.');
+        
         $result = $this->roleService->deleteRole($id);
 
         if ($result['success']) {
